split core space trace and espace trace
