+++ Cargo.toml
@@ -59,7 +59,7 @@ members = [
     "crates/rpc/rpc-cfx-types", 
     "crates/rpc/rpc-eth-api",
     "crates/cfxcore/parity-trace-types",
-    "crates/rpc/rpc",
+    "crates/rpc/rpc", "crates/util/util-macros",
 ]
 
 resolver = "2"
@@ -150,6 +150,7 @@ cfx-rpc-primitives = { path = "./crates/rpc/rpc-primitives" }
 cfx-rpc-cfx-types = { path = "./crates/rpc/rpc-cfx-types" }
 cfx-parity-trace-types = { path = "./crates/cfxcore/parity-trace-types" }
 cfx-rpc-eth-api = { path = "./crates/rpc/rpc-eth-api" }
+cfx-util-macros = { path = "./crates/util/util-macros" }
 
 serde = { version = "1.0", features = ["derive", "alloc"] }
 serde_json = "1.0"
