+++ crates/client/src/rpc/impls/cfx/mod.rs
@@ -4,5 +4,7 @@ pub mod common;
 pub mod light;
 pub mod pool;
 pub mod pubsub;
+mod trace;
 
 pub use cfx_handler::{CfxHandler, LocalRpcImpl, RpcImpl, TestRpcImpl};
+pub use trace::TraceHandler;
