+++ crates/client/src/rpc/impls/eth/mod.rs
@@ -2,6 +2,8 @@ pub mod debug;
 pub mod eth_filter;
 pub mod eth_handler;
 pub mod eth_pubsub;
+pub mod eth_trace;
 
 pub use debug::GethDebugHandler;
 pub use eth_handler::EthHandler;
+pub use eth_trace::EthTraceHandler;
